updated ajna navigation
