+++ components/navigation/NavigationMenuOrb.tsx
@@ -17,10 +17,6 @@ export interface NavigationMenuOrbProps {
   onClose?: () => void
   onOpen?: () => void
   width?: number
-  outline?: {
-    outline: string
-    outlineColor: string
-  }
 }
 
 interface NavigationMenuOrbIconProps {
@@ -43,7 +39,6 @@ export function NavigationOrb({
   onClose,
   onOpen,
   width,
-  outline,
 }: NavigationMenuOrbProps) {
   const [isOpen, toggleIsOpen, setIsOpen] = useToggle(false)
   const didMountRef = useRef(false)
@@ -76,8 +71,6 @@ export function NavigationOrb({
             transition: 'background-color 200ms',
             '&:hover': { boxShadow: 'buttonMenu' },
             ':hover': { color: 'primary100' },
-            ...(isOpen && outline),
-            ':focus': isOpen ? outline : null,
           }}
         >
           <NavigationOrbIcon
